+++ config/configUtils.cpp
@@ -1,69 +1,4 @@
-#include "Utils.hpp"
-
-std::string ft_ltrim(std::string const &str)
-{
-	size_t idx = 0;
-
-	while (isspace(str[idx]))
-		idx++;
-	return str.substr(idx);
-}
-
-std::string ft_rtrim(std::string const &str)
-{
-	size_t idx = str.length() - 1;
-
-	while (isspace(str[idx]))
-		idx--;
-	return str.substr(0, idx + 1);
-}
-
-std::string ft_trim(std::string const &str)
-{
-	std::string ltrim, rtrim;
-
-	ltrim = ft_ltrim(str);
-	rtrim = ft_rtrim(ltrim);
-	return rtrim;
-}
-
-std::vector<std::string> ft_split(std::string const &line, std::string const &delimiter)
-{
-	std::vector<std::string> words;
-	std::string new_line = line;
-	size_t delimiter_len = delimiter.length();
-
-	size_t pos;
-	while ((pos = new_line.find(delimiter)) != std::string::npos) {
-		words.push_back(new_line.substr(0, pos));
-		new_line = new_line.substr(pos + delimiter_len);
-	}
-	words.push_back(new_line);
-
-	return words;
-}
-
-std::vector<std::string> ft_split_space(std::string const &line)
-{
-	std::vector<std::string> words;
-	std::string new_line = line;
-	int i = 0;
-
-	while (i < new_line.length()) {
-		if (isspace(new_line[i])) {
-			if (i > 0)
-				words.push_back(new_line.substr(0, i));
-			while (isspace(new_line[i]))
-				i++;
-			new_line = new_line.substr(i);
-			i = 0;
-		}
-		i++;
-	}
-	if (new_line != "")
-		words.push_back(new_line);
-	return words;
-}
+#include "configUtils.hpp"
 
 int parseSimpleDirective(std::map<std::string, std::string> &directive, const std::string &buffer) {
 	std::vector<std::string> main_line;
